products/biodiversity/requirements.txt
adlfs==2023.12.0
azure-ai-ml==1.12.0
azure-identity==1.15.0
backoff==2.2.1
dask[complete]==2023.5.0
dask-distance==0.2.0
dask-geopandas==0.3.1
dask-mpi==2022.4.0
fsspec==2023.12.2
geojson==3.1.0
geopandas==0.13.2
hyperopt==0.2.7
lightgbm==3.3.2
mlflow==2.9.2
opencensus==0.11.4
opencensus.ext.azure==1.1.13
opencensus-ext-logging==0.1.1
pandas==1.5.3
planetary-computer==1.0.0
pyarrow==14.0.2
pyodbc==5.0.1
pystac==1.8.4
pystac-client==0.7.5
scikit-learn==1.3.2
seaborn==0.13.1
stackstac==0.5.0
utm==0.7.0
xarray==2023.1.0
xarray-spatial==0.3.7
zarr==2.16.1
